requirements.txt
# Formatting & linting (ТЗ: стандарты Python)
black>=24.0
ruff>=0.15

# Plotting experiment CSV (scripts/plot_experiment_csv.py)
matplotlib>=3.7
numpy>=1.24

# Tests (pytest tests/)
pytest>=7.0
PyYAML>=6.0
